backend/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
attrs==24.3.0
build==1.2.2.post1
CacheControl==0.14.1
certifi==2024.12.14
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.8
crashtest==0.4.1
distlib==0.3.9
Django==4.2
django-cors-headers==4.4.0
django-filter==24.2
django-redis==5.4.0
djangorestframework==3.15.2
drf-nested-routers==0.93.4
drf-spectacular==0.28.0
drf-yasg==1.21.8
dulwich==0.21.7
fastjsonschema==2.21.1
filelock==3.16.1
h11==0.14.0
identify==2.6.3
idna==3.10
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jdatetime==4.1.1
keyring==24.3.1
more-itertools==10.5.0
msgpack==1.1.0
nodeenv==1.9.1
packaging==24.2
pexpect==4.9.0
pika==1.3.2
pika-stubs==0.1.3
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
pre_commit==4.0.1
#psycopg2==2.9.10
psycopg2-binary
ptyprocess==0.7.0
pycparser==2.22
pydantic==2.10.3
pydantic_core==2.27.1
PyNaCl==1.5.0
pyproject_hooks==1.2.0
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.11.0
redis==5.0.1
referencing==0.35.1
requests==2.32.3
requests-toolbelt==1.0.0
rest-framework-simplejwt==0.0.2
rpds-py==0.22.3
shellingham==1.5.4
sqlparse==0.5.3
tomlkit==0.13.2
trove-classifiers==2024.10.21.16
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.34.0
virtualenv==20.28.0
xattr==1.1.0
sentry-sdk==1.42.0
boto3==1.40.35
django-storages==1.14.6
pytest==9.0.2
pytest-django==4.11.1
